Added QuestionPanel and TrophicMemberPanel
Clicking on an animal that has been correctly placed now spawns a window
which produces the questions associated with that animal.

If a question has been answered correctly, it will turn green. If all
questions are answered correctly, the image background turns green.

If you click on a trophic level (but not on an icon), it produces a list
of all correctly placed members of that trophic level. Additionally, if
all questions have been answered for a specific animal, its background
appears green.