Saving and Loading now works
Everything is saved when a user answers a question, and on boot, the
game loads from the database. Fixed bugs in.

Added scrollpane to QuestionPanel.

Amended population of database.

Bugfixes regarding titles in the database.

TODO:
Adjust ProfilePanel.
Currently, game does not save "which animals have been placed
correctly." The user will need to re-drag on boot.